Fixed kill, fix background, improved shell

+++ Userland/SampleCodeModule/include/sys_calls.h
@@ -52,7 +52,7 @@ void sys_mem_free(void * ptr);
 
 void * sys_mem_init(int s);
 
-void sys_create_process_set_fd(uint64_t *fd_ids, int fd_count);
+void sys_create_process_set_fd(int *fd_ids, int fd_count);
 
 uint64_t sys_create_process(int priority, program_t program, uint64_t argc, char *argv[]);
